+++ cloudy_salesforce/collections/__init__.py
@@ -9,19 +9,27 @@ from .crud_operations import (
     update,
     upsert,
 )
-from .return_functions import dml_results_only, records_and_response, success_failure
+from .return_functions import (
+    build_dml_results,
+    dml_results_only,
+    records_and_response,
+    success_failure,
+)
+from .types import DmlResult
 
 __all__ = [
     "DeleteProps",
+    "DmlResult",
     "InsertProps",
     "UpdateProps",
     "UpsertProps",
+    "build_dml_results",
     "collections",
     "delete",
-    "insert",
-    "update",
-    "upsert",
     "dml_results_only",
+    "insert",
     "records_and_response",
     "success_failure",
+    "update",
+    "upsert",
 ]